Fetch and store UserInfo while handling token exchange request
This adds the UserInfoInterface and an associated Provider to the
Oauth2Configurator interface so we can pass the dependency down to the
handler. Once in the token exchange handler, the issuer and subject
token are passed to the userinfo service to retreive the information
and to store it in the storage backend.

Signed-off-by: Adam Mohammed <admohammed@equinix.com>